Fixed or disabled warnings in MSVS 2015 (Trail)

+++ src/color_maps.cpp
@@ -634,9 +634,9 @@ namespace Sass {
 
   const char* color_to_name(const Color& c)
   {
-    double key = c.r() * 0x10000
-               + c.g() * 0x100
-               + c.b();
+    int key = (int)(c.r() * 0x10000
+              + c.g() * 0x100
+              + c.b());
     return color_to_name(key);
   }
